+++ packages/react/src/composite/ScrollPane/src/contexts/ScrollPaneContext/ScrollPaneContext.types.tsx
@@ -1,0 +1,8 @@
+import { RefObject } from "react";
+
+export interface ScrollPaneContextValue {
+  direction: "horizontal" | "vertical";
+  scrollToItemOnClick: boolean;
+  containerRef?: RefObject<HTMLDivElement>;
+  scrollToDirection: (direction: "start" | "end") => void;
+}